Nav and Footer now iframes
Changed navbar and footer to iframes so that it makes it easier to edit these and have it effect all pages instead of needing to edit each page individually for these elements.
Current issue is that these will give scrollbars if the window's width is small so this will need to be solved for mobile implementation.